Add live stream lighting mode
The mode handling in main() is quite hacky, but this adds the ability to
listen to an input device and produce lighting commands from that. This
mode is meant to be used with a sound device that has pass-through (or
monitoring) enabled. The audio data is fed to the sound device via
line-in and immediately played back via its low-latency monitoring
feature.
At the same time, this program grabs the input and does its job,
hopefully with reasonably small latency.

Wee! \o/