Extend the 'beat detection' algorithm by some smoothing
The low-pass on the final intensity signal makes it less punchy but also
way more living room compatible. This will certainly need to become
configurable at some point. Buffer size, signal scaling, maybe a
compressor on the input signal?
Color changes as well, but let's keep that for the next round of
commits. ;)