Support monitoring PulseAudio sinks
The existing SDL integration does not support lighting the audio played
on an output device. It can only eat an input device or play a file and
analyze it on the fly. With PulseAudio it is now possible to sniff on
any sink (output device) and have lighting match ... anything!