Fix playing back a single file by ourselves
This broke during the transition to PulseAudio for live analysis because
the first SDL audio device cannot be the default for PA usage.

Now an audio device always needs to be provided. If a local file is also
given for playback, the audio device will be the output device. If no
file is provided, groggle will monitor the given device via PulseAudio.